Refactor

+++ Sources/StableEventDTOUpdate.h
@@ -1,12 +1,14 @@
 #pragma once
 
+#include "../Resources/Orthanc/Plugins/OrthancPluginCppWrapper.h"
+
 struct StableEventDTOUpdate
 {
   /* data */
-  int id_;
+  int64_t id_;
   const char* failed_reason_;
   int retry_;
-  StableEventDTOUpdate(int id, const char* failed_reason, int retry) :
+  StableEventDTOUpdate(int64_t id, const char* failed_reason, int retry) :
       id_(id), failed_reason_(failed_reason), retry_(retry)
   {}
 };
